refactor as useTransfer
