+++ src/components/balance/modals/ModalTransferAmount.vue
@@ -113,7 +113,7 @@
             type="button"
             :disabled="!canExecuteTransaction"
             class="confirm"
-            @click="transfer(transferAmt, fromAddress, toAddress)"
+            @click="transfer"
           >
             {{ $t('confirm') }}
           </button>
@@ -126,24 +126,14 @@
   </div>
 </template>
 <script lang="ts">
-import type { SubmittableExtrinsic, SubmittableExtrinsicFunction } from '@polkadot/api/types';
 import BN from 'bn.js';
 import FormatBalance from 'components/balance/FormatBalance.vue';
 import InputAmount from 'components/common/InputAmount.vue';
 import { getProviderIndex, providerEndpoints } from 'src/config/chainEndpoints';
-import { useChainMetadata, useCustomSignature } from 'src/hooks';
-import { $api } from 'boot/api';
-import {
-  reduceBalanceToDenom,
-  isValidAddressPolkadotAddress,
-  reduceDenomToBalance,
-} from 'src/hooks/helper/plasmUtils';
-import { isValidEvmAddress, toSS58Address, toEvmAddress } from 'src/config/web3/utils/convert';
-import { getUnit } from 'src/hooks/helper/units';
-import { getInjector } from 'src/hooks/helper/wallet';
+import { useChainMetadata, useTransfer } from 'src/hooks';
+import { reduceBalanceToDenom } from 'src/hooks/helper/plasmUtils';
 import { useStore } from 'src/store';
 import { computed, defineComponent, ref, toRefs } from 'vue';
-import Web3 from 'web3';
 import ModalSelectAccount from './ModalSelectAccount.vue';
 
 export enum Role {
@@ -173,10 +163,7 @@ export default defineComponent({
       emit('update:is-open', false);
     };
 
-    const openOption = ref(false);
     const store = useStore();
-    const substrateAccounts = computed(() => store.getters['general/substrateAccounts']);
-
     const { defaultUnitToken, decimal } = useChainMetadata();
 
     const transferAmt = ref<BN>(new BN(0));
@@ -190,8 +177,6 @@ export default defineComponent({
       const chain = chainInfo ? chainInfo.chain : '';
       return getProviderIndex(chain);
     });
-    const isH160 = computed(() => store.getters['general/isH160Formatted']);
-    const { callFunc, handleResult, handleTransactionError } = useCustomSignature(closeModal);
 
     // isCustomSigBlocked is temporary until extrinsic call pallet is deployed to all networks.
     const isCustomSigBlocked = computed(() => !!!providerEndpoints[currentNetworkIdx.value].prefix);
@@ -204,54 +189,15 @@ export default defineComponent({
       return reduceBalanceToDenom(props.accountData.getUsableTransactionBalance(), tokenDecimal);
     });
 
-    const transferLocal = async (transferAmt: BN, fromAddress: string, toAddress: string) => {
-      try {
-        const injector = await getInjector(substrateAccounts.value);
-        const transfer = await $api?.value?.tx.balances.transfer(toAddress, transferAmt);
-        transfer
-          ?.signAndSend(
-            fromAddress,
-            {
-              signer: injector?.signer,
-            },
-            (result) => handleResult(result)
-          )
-          .catch((error: Error) => handleTransactionError(error));
-      } catch (e) {
-        console.error(e);
-      }
-    };
+    const { callTransfer } = useTransfer(selectUnit, decimal, closeModal);
 
-    const transferExtrinsic = async (transferAmt: BN, toAddress: string) => {
-      try {
-        const fn: SubmittableExtrinsicFunction<'promise'> | undefined =
-          $api?.value?.tx.balances.transfer;
-        const method: SubmittableExtrinsic<'promise'> | undefined =
-          fn && fn(toAddress, transferAmt);
-
-        method && callFunc(method);
-      } catch (e) {
-        console.error(e);
-        store.dispatch('general/showAlertMsg', {
-          msg: (e as Error).message,
-          alertType: 'error',
-        });
-      }
-    };
-
-    const transfer = async (transferAmt: number, fromAddress: string, toAddress: string) => {
-      // console.log('transfer', transferAmt);
-      // console.log('fromAccount', fromAddress);
-      // console.log('toAccount', toAddress);
+    const transfer = () => {
+      // console.log('transfer', transferAmt.value);
+      // console.log('fromAccount', fromAddress.value);
+      // console.log('toAccount', toAddress.value);
       // console.log('selUnit', selectUnit.value);
 
-      const toastInvalidAddress = () =>
-        store.dispatch('general/showAlertMsg', {
-          msg: 'balance.modals.invalidAddress',
-          alertType: 'error',
-        });
-
-      if (Number(transferAmt) === 0) {
+      if (Number(transferAmt.value) === 0) {
         store.dispatch('general/showAlertMsg', {
           msg: 'The amount of token to be transmitted must not be zero',
           alertType: 'error',
@@ -259,67 +205,7 @@ export default defineComponent({
         return;
       }
 
-      if (isH160.value) {
-        const provider = typeof window !== 'undefined' && window.ethereum;
-        const web3 = new Web3(provider as any);
-
-        const buildEvmAddress = (toAddress: string) => {
-          // Memo: goes to EVM deposit
-          if (isValidAddressPolkadotAddress(toAddress)) {
-            return toEvmAddress(toAddress);
-          }
-          if (!web3.utils.isAddress(toAddress)) {
-            toastInvalidAddress();
-            return;
-          }
-          return toAddress;
-        };
-
-        const destinationAddress = buildEvmAddress(toAddress);
-
-        store.commit('general/setLoading', true);
-        try {
-          await web3.eth
-            .sendTransaction({
-              to: destinationAddress,
-              from: fromAddress,
-              value: web3.utils.toWei(String(transferAmt), 'ether'),
-            })
-            .once('confirmation', (confNumber, receipt) => {
-              const hash = receipt.transactionHash;
-              const msg = `Completed at transaction hash #${hash}`;
-              store.dispatch('general/showAlertMsg', { msg, alertType: 'success' });
-              store.commit('general/setLoading', false);
-              closeModal();
-            });
-        } catch (error) {
-          console.error(error);
-          store.commit('general/setLoading', false);
-        }
-
-        return;
-      }
-
-      const isValidSS58Address =
-        isValidAddressPolkadotAddress(fromAddress) && isValidAddressPolkadotAddress(toAddress);
-
-      if (!isValidSS58Address && !isValidEvmAddress(toAddress)) {
-        toastInvalidAddress();
-        return;
-      }
-
-      const receivingAddress = isValidEvmAddress(toAddress) ? toSS58Address(toAddress) : toAddress;
-      // console.log('receivingAddress', receivingAddress);
-
-      const unit = getUnit(selectUnit.value);
-      const toAmt = reduceDenomToBalance(transferAmt, unit, decimal.value);
-      // console.log('toAmt', toAmt.toString(10));
-
-      if (isEthWallet.value) {
-        await transferExtrinsic(toAmt, receivingAddress);
-      } else {
-        await transferLocal(toAmt, fromAddress, receivingAddress);
-      }
+      callTransfer(Number(transferAmt.value), fromAddress.value, toAddress.value);
     };
 
     const reloadAmount = (address: string): void => {
@@ -334,7 +220,6 @@ export default defineComponent({
       formatBalance,
       fromAddress,
       toAddress,
-      openOption,
       transferAmt,
       defaultUnitToken,
       selectUnit,
