+++ src/config/web3/utils/index.ts
@@ -57,3 +57,9 @@ export const createWeb3Instance = async (currentNetworkIdx: TNetworkId) => {
 
   return new Web3(new Web3.providers.HttpProvider(network.rpcUrls[0]));
 };
+
+export const getDefaultEthProvider = () => {
+  const provider = typeof window !== 'undefined' && window.ethereum;
+  const web3 = new Web3(provider as any);
+  return web3;
+};
